Moved ITextExtractor interface to seperate file
