+++ src/TikaOnDotnet.TextExtractor/ITextExtractor.cs
@@ -1,0 +1,41 @@
+using System;
+using java.io;
+using org.apache.tika.metadata;
+
+namespace TikaOnDotNet.TextExtraction
+{
+    public interface ITextExtractor
+    {
+        /// <summary>
+        ///     Extract the text from the given file and returns it as an <see cref="TextExtractionResult" /> object
+        /// </summary>
+        /// <param name="filePath">The file with its full path</param>
+        /// <returns></returns>
+        /// <exception cref="TextExtractionException"></exception>
+        TextExtractionResult Extract(string filePath);
+
+        /// <summary>
+        ///     Extract the text from the given byte array and returns it as an <see cref="TextExtractionResult" /> object
+        /// </summary>
+        /// <param name="data">The byte array</param>
+        /// <returns></returns>
+        /// <exception cref="TextExtractionException"></exception>
+        TextExtractionResult Extract(byte[] data);
+
+        /// <summary>
+        ///     Extract from the give uri and returns it as an <see cref="TextExtractionResult" /> object
+        /// </summary>
+        /// <param name="uri"></param>
+        /// <returns></returns>
+        /// <exception cref="TextExtractionException"></exception>
+        TextExtractionResult Extract(Uri uri);
+
+        /// <summary>
+        ///     Extract the text from the given inputstreams and returns it as an <see cref="TextExtractionResult" /> object
+        /// </summary>
+        /// <param name="streamFactory"></param>
+        /// <returns></returns>
+        /// <exception cref="TextExtractionException"></exception>
+        TextExtractionResult Extract(Func<Metadata, InputStream> streamFactory);
+    }
+}
