+++ src/TikaOnDotnet.TextExtractor/TikaOnDotnet.TextExtraction.csproj
@@ -46,6 +46,7 @@
     <Compile Include="..\..\SolutionInfo.cs">
       <Link>Properties\SolutionInfo.cs</Link>
     </Compile>
+    <Compile Include="ITextExtractor.cs" />
     <Compile Include="MySystemClassLoader.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StringExtensions.cs" />
